[2.8] MOD-8275: Add A Function To Format A Parsed Expression (#5674)
* MOD-8275: Add A Function To Format A Parsed Expression (#5456)

* * initial commit

* * Fix Code Review Comments

* * Make standalone dump sds test

* * code review comment

* * fix testDump test

(cherry picked from commit 6df921265228295c921acb62fae614233611587b)

* * fix test

+++ src/obfuscation/obfuscation_api.c
@@ -24,7 +24,7 @@ char *Obfuscate_Text(const char* text) {
   return "Text";
 }
 
-char *Obfuscate_Number(size_t number) {
+char *Obfuscate_Number(double number) {
   return "Number";
 }
